repo-updater: Move OtherSource to other.go
